requirements.txt
streamlit
pandas
pillow
google-generativeai
google-cloud-vision
xlsxwriter

#flask


